+++ src/utils/spawn.ts
@@ -1,0 +1,10 @@
+import { handleFfmpegError } from './error-handler'
+
+export function withSpawnError<T>(fn: () => T): T {
+  try {
+    return fn()
+  }
+  catch (error) {
+    throw handleFfmpegError(error)
+  }
+}